Crate json_tools

For usage examples, please have a look at the tests and benchmarks.

Structs

FilterTypedKeyValuePairs
Removes tokens matching ,? "key": <type> ,?., where <type> is a given token type. Useful for removing null values, or all numbers, for instance. Is made in a resilient fashion which doesn't require a sane input token stream.

Lexer
A lexer for utf-8 encoded json data

Span
A pair of indices into the byte stream returned by our source iterator. It is an exclusive range.

Token
A lexical token, identifying its kind and span.

TokenReader
An adapter to convert a stream of Tokens into bytes by implementing an std::io::Read trait.

Enums

Buffer
Representation of a buffer containing items making up a Token.

BufferType
The type of Buffer you want in each Token

TokenType

Traits

IteratorExt
Applies convenience constructors to all Iterator<Item=Token> types