Enum json_tools::Buffer

pub enum Buffer {
    MultiByte(Vec<u8>),
    Span(Span),
}

Representation of a buffer containing items making up a Token.

It's either always Span, or one of the *Byte variants.

Variants

MultiByte

Multiple bytes making up a token. Only set for TokenType::String and TokenType::Number.

Span

The span allows to reference back into the source byte stream to obtain the string making up the token. Please note that for control characters, booleans and null (i.e anything that is not Buffer::MultiByte you should use <TokenType as AsRef<str>>::as_ref())